Fix Firebase error and add React Hot Toast notifications for list creation
