+++ package.json
@@ -10,6 +10,7 @@
 		"firebase": "^10.12.5",
 		"react": "^18.3.1",
 		"react-dom": "^18.3.1",
+		"react-hot-toast": "^2.4.1",
 		"react-router-dom": "^6.26.0"
 	},
 	"devDependencies": {
